Added reconnecting websockets for auto reconnect after unlocting phone.
