+++ include/TFT_AP.h
@@ -1,0 +1,6 @@
+
+void status(const char *msg);
+
+void TFT_AP_setup();
+
+void TFT_AP_loop();
