+++ platformio.ini
@@ -20,6 +20,8 @@ lib_deps =
 	poelstra/MultiButton@^1.2.0
 	robtillaart/HeartBeat@^0.3.2
 	bodmer/TFT_eSPI@^2.5.23
+  ;https://github.com/nagyist/reconnecting-websocket
+
 monitor_speed = 115200
 
 build_flags =
